Launch SuperSU from Settings
If SuperSU is installed, show an entry in Settings below
"Developer options", which launches SuperSU when tapped.
At boot time SuperSU will detect if com.android.settings
contains the resources from this patch, and if so it
will hide itself from the launcher. This was implemented
in v1.85, so we only show the entry if the installed
SuperSU version is at least that.

Conflicts:
	src/com/android/settings/SettingsActivity.java

	new file:   res/drawable-hdpi/ic_settings_supersu.png
	new file:   res/drawable-mdpi/ic_settings_supersu.png
	new file:   res/drawable-xhdpi/ic_settings_supersu.png
	new file:   res/drawable/ic_settings_supersu_teal.xml
	modified:   res/values/liquid_strings.xml
	modified:   res/xml/dashboard_categories.xml
	modified:   src/com/android/settings/SettingsActivity.java

Change-Id: Ib5ed28203a3a1e3e1f54689c9a8a3b74e0917ed1

Conflicts:
	res/values/liquid_strings.xml